AdoPet.Testes/HttpClientPetTest.cs
using Adopet.Console.Servicos;

namespace Adopet.Testes
{
    public class HttpClientPetTest
    {
        [Fact]
        public async Task ListaPetsDeveRetornarUmaListaNaoVazia()
        {
            //Arrange
            var clientePet = new HttpClientPet();

            //Act
            var lista = await clientePet.ListPetsAsync();

            //Assert
            Assert.NotNull(lista);
            Assert.NotEmpty(lista);
        }

        [Fact]
        public async Task QuandoAPIForaDeveRetornarUmaExcecao()
        {
            //Arrange
            var clientePet = new HttpClientPet(uri: "http://localhost:1111");

            //Act+Assert
            await Assert.ThrowsAnyAsync<Exception>(() => clientePet.ListPetsAsync());

        }
    }
}